refactor: flatten lenis component
